feat: Update createChatRoom mutation to accept an array of user IDs
The code changes in this commit modify the createChatRoom mutation in the api.ts file to accept an array of user IDs instead of a single string. This allows for creating a chat room with multiple users. The mutation now takes an array of [myId, userId] and sends a POST request to the /rooms endpoint with the necessary parameters. This change improves the functionality of creating chat rooms in the application.

Note: This commit message follows the established commit message conventions in the repository.